# backend/api-doc.yaml
openapi: 3.0.0
info:
  title: Kulhad Flask Backend API
  version: 1.0.0
  description: |
    OpenAPI 3.0 specification generated from the current Flask backend blueprints.

    Authentication-protected endpoints use JWT bearer authentication.
    The computer vision endpoints upload an image to the backend, which then calls an external object-detection service internally.
servers:
  - url: http://localhost:5000
    description: Local development server
...
paths:
...
  /api/products/:
    get:
      tags:
        - products
      summary: List products
      description: Returns the full product catalog. This endpoint is public.
      responses:
        '200':
          description: Product list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
              example:
                - id: 1
                  name: Classic Kulhad
                  category: Kulhad
                  price: 20.0
                  wage_per_kulhad: 25.0
                  stock: 120
                  image: classic.png
    post:
      tags:
        - products
      summary: Create a product
      description: Creates a new catalog product. Admin access is required.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCreateRequest'
            example:
              name: Premium Serving Kulhad
              category: Kulhad
              price: 65
              wage_per_kulhad: 40
              stock: 18
              image: premium-serving-kulhad.webp
      responses:
        '201':
          description: Product created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                message: Product created
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/products/images/{filename}:
    get:
      tags:
        - products
      summary: Fetch a product image
      description: Serves a static product image file by filename.
      parameters:
        - $ref: '#/components/parameters/Filename'
      responses:
        '200':
          description: Image file response
          content:
            image/png: {}
            image/jpeg: {}
            image/webp: {}
        '404':
          description: Image not found
  /api/products/upload-image:
    post:
      tags:
        - products
      summary: Upload a product image
      description: Uploads a product image into backend static storage. Admin access is required.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - image
              properties:
                image:
                  type: string
                  format: binary
            encoding:
              image:
                contentType: image/png, image/jpeg, image/webp
      responses:
        '201':
          description: Image uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductImageUploadResponse'
              example:
                filename: upload.jpg
        '400':
          description: Missing or unsupported image
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                missing:
                  value:
                    error: Image file is required
                invalid:
                  value:
                    error: Only PNG, JPG, JPEG, and WEBP images are supported
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/products/history:
    get:
      tags:
        - products
      summary: List product history
      description: Returns the latest 100 product history entries. Admin access is required.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Product history list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductHistoryEntry'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/products/{product_id}:
    put:
      tags:
        - products
      summary: Update a product
      description: Updates one or more product fields. Admin access is required.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProductId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductUpdateRequest'
            example:
              price: 19.0
              stock: 99
              image: updated.jpg
      responses:
        '200':
          description: Product updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                message: Product updated
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/ProductNotFound'
    delete:
      tags:
        - products
      summary: Delete a product
      description: Deletes a product. Admin access is required.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProductId'
      responses:
        '200':
          description: Product deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                message: Product deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/ProductNotFound'
...
  /api/production/entries:
    get:
      tags:
        - production
      summary: Get employee production entries
      description: Returns production attendance rows for the authenticated employee.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Month'
        - $ref: '#/components/parameters/Year'
      responses:
        '200':
          description: Production entry list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductionEntry'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Non-employee account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Only employees can access production entries
        '404':
          $ref: '#/components/responses/UserNotFound'
  /api/production/log:
    post:
      tags:
        - production
      summary: Log production
      description: Creates a production entry for the authenticated employee.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductionLogRequest'
            example:
              date: '2026-04-20'
              product_id: 1
              product: Classic Kulhad
              quantity: 8
              method: manual
              entry_type: manual
              status: present
              notes: Manual logging
      responses:
        '201':
          description: Production entry created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductionEntry'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Non-employee account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Only employees can access production entries
        '404':
          $ref: '#/components/responses/UserNotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ProductId:
      name: product_id
      in: path
      required: true
      schema:
        type: integer
      description: Product identifier.
    Filename:
      name: filename
      in: path
      required: true
      schema:
        type: string
      description: Static image filename.
...
    Month:
      name: month
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 12
      description: Calendar month.
    Year:
      name: year
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
      description: Calendar year.
...
  responses:
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AuthErrorResponse'
          example:
            msg: Missing Authorization Header
    Forbidden:
      description: Authenticated user does not have the required role
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/MessageResponse'
          example:
            message: Access forbidden
    Error400:
      description: Bad request
      content:
        application/json:
          schema:
            oneOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - $ref: '#/components/schemas/MessageResponse'
    UserNotFound:
      description: User not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: User not found
    ProductNotFound:
      description: Product not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Product not found
...
  schemas:
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    AuthErrorResponse:
      type: object
      properties:
        msg:
          type: string
      required:
        - msg
...
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        price:
          type: number
          format: float
        wage_per_kulhad:
          type: number
          format: float
        stock:
          type: integer
        image:
          type: string
      required:
        - id
        - name
        - category
        - price
        - wage_per_kulhad
        - stock
        - image
      example:
        id: 1
        name: Classic Kulhad
        category: Kulhad
        price: 20.0
        wage_per_kulhad: 25.0
        stock: 120
        image: classic.png
    ProductCreateRequest:
      type: object
      required:
        - name
        - category
        - price
        - stock
        - wage_per_kulhad
      properties:
        name:
          type: string
        category:
          type: string
        price:
          type: number
          format: float
        wage_per_kulhad:
          type: number
          format: float
        stock:
          type: integer
        image:
          type: string
    ProductUpdateRequest:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        price:
          type: number
          format: float
        wage_per_kulhad:
          type: number
          format: float
        stock:
          type: integer
        image:
          type: string
    ProductImageUploadResponse:
      type: object
      properties:
        filename:
          type: string
      required:
        - filename
    ProductHistoryEntry:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        desc:
          type: string
        date:
          type: string
        by:
          type: string
        type:
          type: string
      required:
        - id
        - name
        - desc
        - date
        - by
        - type
...
    ProductionEntry:
      type: object
      properties:
        id:
          type: integer
        product_id:
          type: integer
          nullable: true
        product_name:
          type: string
        date:
          type: string
          format: date
        status:
          type: string
        type:
          type: string
        kulhad_quantity:
          type: integer
        cup_quantity:
          type: integer
        total_quantity:
          type: integer
        wage_per_kulhad:
          type: number
          format: float
        daily_wage:
          type: number
          format: float
        notes:
          type: string
        created_at:
          type: string
      required:
        - id
        - product_name
        - date
        - status
        - type
        - kulhad_quantity
        - cup_quantity
        - total_quantity
        - wage_per_kulhad
        - daily_wage
        - notes
        - created_at
    ProductionLogRequest:
      type: object
      properties:
        date:
          type: string
          format: date
        product_id:
          type: integer
        product:
          type: string
        quantity:
          type: integer
        kulhad_quantity:
          type: integer
        method:
          type: string
          enum:
            - manual
            - regular
            - overtime
            - image
            - leave
        entry_type:
          type: string
          enum:
            - regular
            - overtime
            - manual
            - image
            - leave
        status:
          type: string
          enum:
            - present
            - leave
        notes:
          type: string
...